pyproject.toml
[project]
name = "nfway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "geopy>=2.4.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "python-telegram-bot>=22.3",
    "qreader>=3.16",
    "requests>=2.32.5",
]

[project.scripts]
nfway = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "matplotlib>=3.10.6",
    "pyside6>=6.9.2",
    "ruff>=0.12.12",
]
